requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==1.11.3
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.3.3
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
psycopg2==2.8.4
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
